Move TreeTables setup into tree/index.jsx
