+++ templates/codelists/version.html
@@ -339,5 +339,5 @@
 {{ code_to_term|json_script:"code-to-term" }}
 {{ code_to_status|json_script:"code-to-status" }}
 
-<script src="{% static 'js/tree-tables.bundle.js' %}"></script>
+<script src="{% static 'js/tree.bundle.js' %}"></script>
 {% endblock %}
